Restructure news scrapers into `news` package
In order to make a big set of scrapers more manageable, this creates a `news` package where we can have one module per scraper (plus other support modules). The same script as before is the entry point, but you specify counties as arguments, e.g:

    > python3 scraper_news.py san_francisco

If you don't specify any arguments it currently defaults to `san_francisco`, but this will eventually be turned into a combined feed of all counties instead.
