+++ news/__init__.py
@@ -1,0 +1,14 @@
+from .san_francisco import SanFranciscoNews
+
+
+scrapers = {
+    'alameda': None,
+    'contra_costa': None,
+    'marin': None,
+    'napa': None,
+    'san_francisco': SanFranciscoNews,
+    'san_mateo': None,
+    'santa_clara': None,
+    'solano': None,
+    'sonoma': None,
+}
